typo

+++ wifi/install.sh
@@ -24,7 +24,7 @@ bw list --folderid $FOLDERID items > items.json
 # remove files from /etc
 sudo rm -f /etc/NetworkManager/system-connections/$FILEPREFIX*$FILESUFFIX
 
-[ if -e $STAGING/$FILEPREFIX*$FILESUFFIX ];
+if [ -e $STAGING/$FILEPREFIX*$FILESUFFIX ];
 then
     # copy files from staging
     sudo cp $STAGING/$FILEPREFIX*$FILESUFFIX /etc/NetworkManager/system-connections/
